+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Changelog
+
+* 2.4.0
+  - supports [new syntax](https://playwright.dev/docs/test-annotations) for tag and annotation in playwright v1.42.0
+
 * 2.3.4
   - fixed dependency `open` to support Node 16
   - updated `mcr` version to support `all` for adding empty coverage for all files